requirements.txt
flask
google-generativeai
python-dotenv
openpyxl
faker